gateway/requirement.txt
Django==5.1.4 
djangorestframework==3.15.2 
django-environ==0.11.2 
